# src/main/resources/api-docs.yaml
openapi: 3.1.0
info:
  title: OpenAPI definition
  version: v0
servers:
- url: http://localhost:8080
  description: Generated server url
paths:
  /restaurante/platos/{id}:
    get:
      tags:
      - platos-controller-impl
      summary: Obtener plato a partir del ID
      operationId: getPlatosById
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/Plato"
    put:
      tags:
      - platos-controller-impl
      summary: Actualizar plato
      operationId: updatePlato
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Plato"
        required: true
      responses:
        "200":
          description: OK
    delete:
      tags:
      - platos-controller-impl
      summary: Borrar plato
      operationId: deletePlato
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: OK
  /restaurante/locales/{cif}:
    get:
      tags:
      - locales-controller-impl
      summary: Obtener local a partir de CIF
      operationId: getLocalByCif
      parameters:
      - name: cif
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/Local"
    put:
      tags:
      - locales-controller-impl
      summary: Crear o actualizar local
      operationId: createOrUpdateLocal
      parameters:
      - name: cif
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Local"
        required: true
      responses:
        "200":
          description: OK
    delete:
      tags:
      - locales-controller-impl
      summary: Borrar local
      operationId: deleteLocal
      parameters:
      - name: cif
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
  /restaurante/platos:
    get:
      tags:
      - platos-controller-impl
      summary: Obtener lista de platos
      operationId: getPlatos
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Plato"
    post:
      tags:
      - platos-controller-impl
      summary: Crear plato
      operationId: createPlato
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Plato"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: integer
                format: int64
  /restaurante/locales/{cif}/platos:
    get:
      tags:
      - locales-controller-impl
      summary: Obtener lista del menu platos del local
      operationId: getPlatos_1
      parameters:
      - name: cif
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Plato"
    post:
      tags:
      - locales-controller-impl
      summary: Añadir plato al menu del local
      operationId: addPlato
      parameters:
      - name: cif
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              type: integer
              format: int32
        required: true
      responses:
        "200":
          description: OK
  /restaurante/locales:
    get:
      tags:
      - locales-controller-impl
      summary: Obtener lista de locales
      operationId: getLocales
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Local"
  /restaurante/locales/{cif}/platos/{plato}:
    delete:
      tags:
      - locales-controller-impl
      summary: Borrar plato del menu del local
      operationId: removePlato
      parameters:
      - name: cif
        in: path
        required: true
        schema:
          type: string
      - name: plato
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: OK
components:
  schemas:
    Plato:
      type: object
      properties:
        id:
          type: integer
          format: int32
        nombre:
          type: string
        precio:
          type: number
          format: double
        categoria:
          type: string
          enum:
          - Entrante
          - Plato principal
          - Postre
    Local:
      type: object
      properties:
        cif:
          type: string
          description: CIF del local
        nombre:
          type: string
          description: Nombre del local
        direccion:
          type: string
          description: Dirección del local
        telefono:
          type: string
          description: Teléfono del local
        carta:
          type: array
          description: Lista de cartas del local
          items:
            $ref: "#/components/schemas/Plato"